--- a/rust/.cargo/config.toml
+++ b/rust/.cargo/config.toml
@@ -1,0 +1,5 @@
+[build]
+rustflags = [
+  "-C",
+  "target-cpu=native",
+]