--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -3,11 +3,9 @@ name = "rust"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-ahash = "0.8"
 aligned-vec = "0.5"
+hashbrown = "0.14.3"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"