--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -2,6 +2,7 @@
 name = "rust"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [dependencies]
 aligned-vec = "0.5"