--- a/rust/src/main.rs
+++ b/rust/src/main.rs
@@ -3,7 +3,6 @@ use hashbrown::HashMap;
 use serde::{Deserialize, Serialize};
 use std::{
     fs::OpenOptions,
-    hint,
     io::{self, BufWriter},
     time::Instant,
 };
@@ -94,10 +93,8 @@ fn main() -> io::Result<()> {
         })
         .collect();
 
-    // Tell compiler to not delay now() until print is eval'ed.
-    let end = hint::black_box(Instant::now());
-
-    println!("Processing time (w/o IO): {:?}", end.duration_since(start));
+    let elapsed = start.elapsed();
+    println!("Processing time (w/o IO): {elapsed:?}");
 
     let output_file = OpenOptions::new()
         .write(true)