--- a/rust_con/Cargo.toml
+++ b/rust_con/Cargo.toml
@@ -2,8 +2,7 @@
 name = "rust_rayon"
 version = "0.1.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+publish = false
 
 [dependencies]
 mimalloc = "0.1.39"