--- a/rust_con/Cargo.toml
+++ b/rust_con/Cargo.toml
@@ -13,7 +13,7 @@ serde = { version = "1.0.189", features = ["derive"] }
 serde_json = "1.0.107"
 
 [profile.release]
-lto = "thin"
+lto = "fat"
 codegen-units = 1
 panic = "abort"
 debug = false