I updated dependencies, cleaned up some code and most importantly: I improved the overall performance by writing JSON directly into the output file.