@jinyus 2024-01-17
Single core version is slower for the 2nd and 3rd rounds. Confirmed with 2 extra runs.
Multicore is faster

Main:
```
| Rust | 10.24 ms | 123.20 ms | 1.06 s | 1.19 s |
```

This PR:
```yaml
Rust:

        Processing time (w/o IO): 10.765912ms
        total: 0.02s memory: 8196k
        Processing time (w/o IO): 10.720821ms
        total: 0.05s memory: 8192k
        Processing time (w/o IO): 10.900933ms
        total: 0.06s memory: 8320k
        Processing time (w/o IO): 10.708752ms
        total: 0.03s memory: 8188k
        Processing time (w/o IO): 10.744661ms
        total: 0.06s memory: 8192k
        Processing time (w/o IO): 10.56677ms
        total: 0.07s memory: 8192k
        Processing time (w/o IO): 10.708381ms
        total: 0.06s memory: 8320k
        Processing time (w/o IO): 10.968762ms
        total: 0.02s memory: 8196k
        Processing time (w/o IO): 10.66134ms
        total: 0.02s memory: 8064k
        Processing time (w/o IO): 10.744031ms
        total: 0.02s memory: 8320k

Rust:

        Processing time (w/o IO): 133.707362ms
        total: 0.33s memory: 25084k
        Processing time (w/o IO): 134.218785ms
        total: 0.32s memory: 25216k
        Processing time (w/o IO): 133.689503ms
        total: 0.34s memory: 25220k

Rust:

        Processing time (w/o IO): 1.115432331s
        total: 1.83s memory: 70272k
        Processing time (w/o IO): 1.101599102s
        total: 1.41s memory: 70404k
        Processing time (w/o IO): 1.110922934s
        total: 1.75s memory: 70272k
```

Main:
```
| Rust Concurrent | 4.71 ms | 60.90 ms | 491.75 ms | 557.36 ms |
```

This PR:
```yaml
Rust Concurrent:

        Processing time (w/o IO): 4.211594ms
        total: 0.01s memory: 13312k
        Processing time (w/o IO): 4.200883ms
        total: 0.02s memory: 13312k
        Processing time (w/o IO): 4.349095ms
        total: 0.02s memory: 13440k
        Processing time (w/o IO): 4.182563ms
        total: 0.02s memory: 13440k
        Processing time (w/o IO): 4.612766ms
        total: 0.02s memory: 13568k
        Processing time (w/o IO): 4.243944ms
        total: 0.02s memory: 13568k
        Processing time (w/o IO): 4.484835ms
        total: 0.02s memory: 13440k
        Processing time (w/o IO): 4.246384ms
        total: 0.02s memory: 13312k
        Processing time (w/o IO): 4.483735ms
        total: 0.02s memory: 13568k
        Processing time (w/o IO): 4.308644ms
        total: 0.01s memory: 13440k

Rust Concurrent:

        Processing time (w/o IO): 56.99422ms
        total: 0.13s memory: 36168k
        Processing time (w/o IO): 57.826444ms
        total: 0.13s memory: 36292k
        Processing time (w/o IO): 58.097856ms
        total: 0.13s memory: 36404k

Rust Concurrent:

        Processing time (w/o IO): 487.223735ms
        total: 0.75s memory: 108396k
        Processing time (w/o IO): 494.657137ms
        total: 0.93s memory: 108428k
        Processing time (w/o IO): 486.974385ms
        total: 1.23s memory: 108604k
```