@mo8it 2024-01-17
I also noticed the speedup of the concurrent version. But I can't explain it, to be honest. Because the change should be outside the benchmarked code.

I would like to debug the regression in the sequential version. How do you run these rounds? Running `run.sh rust` run only the first round.