@jinyus 2024-01-17
I generate fake data between runs. You need faker and nanoid installed for python.
```
pip install faker nanoid
```

```
 python gen_fake_posts.py 20000
./run.sh rust 

python gen_fake_posts.py 60000
./run.sh rust
```