@mo8it 2024-01-17
I found out how to generate a file with 60k posts by using `gen_fake_posts.py 60000`. But I can't reproduce a regression for the sequential version by comparing to the main branch. It is in the range of normal fluctuation on my machine :/